Local coverage is resolved
Revert to old changes
