+++ api/src/test/java/uk/gov/hmcts/payment/api/v1/componenttests/PaymentsRefDataComponentTest.java
@@ -1,153 +1,66 @@
 package uk.gov.hmcts.payment.api.v1.componenttests;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.github.tomakehurst.wiremock.junit.WireMockClassRule;
-import org.junit.ClassRule;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit4.SpringRunner;
-import uk.gov.hmcts.payment.api.controllers.PaymentReferenceDataController;
-import uk.gov.hmcts.payment.api.model.LegacySite;
-import uk.gov.hmcts.payment.api.model.LegacySiteRepository;
 import uk.gov.hmcts.payment.api.model.PaymentChannel;
-import uk.gov.hmcts.payment.api.model.PaymentChannelRepository;
 import uk.gov.hmcts.payment.api.model.PaymentMethod;
-import uk.gov.hmcts.payment.api.model.PaymentMethodRepository;
 import uk.gov.hmcts.payment.api.model.PaymentProvider;
-import uk.gov.hmcts.payment.api.model.PaymentProviderRepository;
 import uk.gov.hmcts.payment.api.model.PaymentStatus;
-import uk.gov.hmcts.payment.api.model.PaymentStatusRepository;
-import uk.gov.hmcts.payment.api.v1.componenttests.backdoors.UserResolverBackdoor;
 
-import java.util.Collections;
-import java.util.List;
-
-import static org.junit.Assert.assertEquals;
-import static org.mockito.Mockito.when;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 
-@RunWith(SpringRunner.class)
-@ActiveProfiles({"local", "componenttest"})
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.MOCK)
-class PaymentRefDataComponentTest {
-
-    private static final String USER_ID = UserResolverBackdoor.AUTHENTICATED_USER_ID;
-    @ClassRule
-    public static WireMockClassRule wireMockRule = new WireMockClassRule(9190);
-    @Autowired
-    private ObjectMapper objectMapper;
-    @InjectMocks
-    private PaymentReferenceDataController paymentReferenceDataController;
-    @Mock
-    private LegacySiteRepository legacySiteRepository;
-    @Mock
-    private PaymentStatusRepository paymentStatusRepository;
-    @Mock
-    private PaymentProviderRepository paymentProviderRepository;
-    @Mock
-    private PaymentMethodRepository paymentMethodRepository;
-    @Mock
-    private PaymentChannelRepository paymentChannelRepository;
+public class PaymentsRefDataComponentTest extends PaymentsComponentTest {
 
     @Test
     public void testFindAllPaymentChannels() throws Exception {
-//
-//        restActions
-//            .get("/refdata/channels")
-//            .andExpect(status().isOk())
-//            .andExpect(body().asListOf(PaymentChannel.class, paymentChannels -> {
-//                assertThat(paymentChannels).anySatisfy(paymentChannel -> {
-//                    assertThat(paymentChannel.getName()).isEqualTo("Online");
-//                });
-//            }));
 
-        //Controller Mock Test
-        List<PaymentChannel> paymentChannelMock = Collections.singletonList(new PaymentChannel("card", "card payment"));
-        when(paymentChannelRepository.findAll()).thenReturn(paymentChannelMock);
-        List<PaymentChannel> paymentChannelMockResponse = paymentReferenceDataController.findAllPaymentChannels();
-        assertEquals(paymentChannelMock, paymentChannelMockResponse);
+        restActions
+            .get("/refdata/channels")
+            .andExpect(status().isOk())
+            .andExpect(body().asListOf(PaymentChannel.class, paymentChannels -> {
+                assertThat(paymentChannels).anySatisfy(paymentChannel -> {
+                    assertThat(paymentChannel.getName()).isEqualTo("Online");
+                });
+            }));
     }
 
     @Test
     public void testFindAllPaymentMethods() throws Exception {
 
-//        restActions
-//            .get("/refdata/methods")
-//            .andExpect(status().isOk())
-//            .andExpect(body().asListOf(PaymentMethod.class, paymentMethods -> {
-//                assertThat(paymentMethods).anySatisfy(paymentMethod -> {
-//                    assertThat(paymentMethod.getName()).isEqualTo("Card");
-//                });
-//            }));
-
-        //Controller Mock Test
-        List<PaymentMethod> paymentMethodMock = Collections.singletonList(new PaymentMethod("online", "card payment"));
-        when(paymentMethodRepository.findAll()).thenReturn(paymentMethodMock);
-        List<PaymentMethod> paymentMethodMockResponse = paymentReferenceDataController.findAllPaymentMethods();
-        assertEquals(paymentMethodMock, paymentMethodMockResponse);
+        restActions
+            .get("/refdata/methods")
+            .andExpect(status().isOk())
+            .andExpect(body().asListOf(PaymentMethod.class, paymentMethods -> {
+                assertThat(paymentMethods).anySatisfy(paymentMethod -> {
+                    assertThat(paymentMethod.getName()).isEqualTo("Card");
+                });
+            }));
     }
 
     @Test
     public void testFindAllPaymentProviders() throws Exception {
 
-//        restActions
-//            .get("/refdata/providers")
-//            .andExpect(status().isOk())
-//            .andExpect(body().asListOf(PaymentProvider.class, paymentProviders -> {
-//                assertThat(paymentProviders).anySatisfy(paymentProvider -> {
-//                    assertThat(paymentProvider.getName()).isEqualTo("GovPay");
-//                });
-//            }));
-        //Controller Mock Test
-        List<PaymentProvider> paymentProviderMock = Collections.singletonList(new PaymentProvider().GOV_PAY);
-        when(paymentProviderRepository.findAll()).thenReturn(paymentProviderMock);
-        List<PaymentProvider> paymentProviderMockResponse = paymentReferenceDataController.findAllPaymentProviders();
-        assertEquals(paymentProviderMock, paymentProviderMockResponse);
+        restActions
+            .get("/refdata/providers")
+            .andExpect(status().isOk())
+            .andExpect(body().asListOf(PaymentProvider.class, paymentProviders -> {
+                assertThat(paymentProviders).anySatisfy(paymentProvider -> {
+                    assertThat(paymentProvider.getName()).isEqualTo("GovPay");
+                });
+            }));
     }
 
     @Test
     public void testFindAllPaymentStatuses() throws Exception {
 
-//        MvcResult mvcResult = restActions
-//            .get("/refdata/status")
-//            .andExpect(status().isOk())
-//            .andReturn();
-//        List<PaymentStatus> paymentStatuses = objectMapper.readValue(mvcResult.getResponse().getContentAsString(), new TypeReference<List<PaymentStatus>>() {
-//        });
-//        PaymentStatus paymentStatus = paymentStatuses.get(0);
-//        assertEquals(paymentStatus.getName(), "created");
-        //Controller Mock Test
-        List<PaymentStatus> paymentStatusMock = Collections.singletonList(new PaymentStatus().CREATED);
-        when(paymentStatusRepository.findAll()).thenReturn(paymentStatusMock);
-        List<PaymentStatus> paymentStatusMockResponse = paymentReferenceDataController.findAllPaymentStatuses();
-        assertEquals(paymentStatusMock, paymentStatusMockResponse);
+        restActions
+            .get("/refdata/status")
+            .andExpect(status().isOk())
+            .andExpect(body().asListOf(PaymentStatus.class, paymentStatuses -> {
+                assertThat(paymentStatuses).anySatisfy(paymentStatus -> {
+                    assertThat(paymentStatus.getName()).isEqualTo("Pending");
+                });
+            }));
     }
-
-
-    @Test
-    public void testFindAllLegacySites() throws Exception {
-//        MvcResult mvcResult = restActions
-//            .withAuthorizedUser(USER_ID)
-//            .withUserId(USER_ID)
-//            .withAuthorizedService("divorce")
-//            .get("/refdata/legacy-sites")
-//            .andExpect(status().isOk())
-//            .andReturn();
-//        List<LegacySite> legacySites = objectMapper.readValue(mvcResult.getResponse().getContentAsString(), new TypeReference<List<LegacySite>>() {
-//        });
-//        LegacySite legacySite = legacySites.get(0);
-//        assertEquals(legacySite.getSiteId(), "Y402");
-//        assertEquals(legacySite.getSiteName(), "Aberdare County Court");
-        //Controller Mock Test
-        List<LegacySite> legacySiteMock = Collections.singletonList(new LegacySite("site 2", "Site name 1"));
-        when(legacySiteRepository.findAll()).thenReturn(legacySiteMock);
-        List<LegacySite> legacySiteMockResponse = paymentReferenceDataController.findAllLegacySites();
-        assertEquals(legacySiteMock, legacySiteMockResponse);
-    }
-
 }
